Made the request tickets anonymous in the ProviderListener flow.
Previously the ProviderListener was identifying the requests by
the Filter that defined them. As of this commit the tickets are
rather passed on as Objects as of the interface and it's then up
to the implementation to implement a ticket management system
(which very well may be the filter anyway).