Made the ItemParser a bit softer on default JSON return value.
Previously the ItemParser fell back on a hard null pointer when trying, and failing, to
convert an Object to a JSON string. That causes a huge set of unnecessary null
checks throughout the code base. As of this change the ItemParser falls back to
an empty string ("") which won't cause unnecessary NullPointerExceptions.