Refactored to SDK concurrency model.
The SDK now returns a Future<RestResult<?>> when enquing a new request.
This Future can then be used to block the current thread until a result is
delivered, hence offering a synchronized behavior (even though the actual
request execution is performed on a separate worker thread). The Future
also renders the previous "ticket" concept invalid, which has therefore been
removed.