Moved the Podio SDK callback injection to the request Future.
The old implementation took the callback implementations as
arguments to the `Provider` methods. These were unconditional
arguments, and the caller had to provide a value for them, even
so if just a null pointer.

The new implementation allows the caller to attach the callback
implementations of interest directly on the returned `Future`
from the `Provider` methods. This is a much more convenient way
of using a modern SDK and enables much more fluent and self
describing source code.