Temporary quick-n-dirty fix for unhandled exception issue.
Currently the SDK delivers all and every failures through an Exception.
The "crash" described in the related issue is what seems to be the client
failing in handeling such a failure. It is, hence, not a crash as in a
malfunctioning SDK, but rather a result of an incomplete implementation
in the client.

Fixes: #2983