Introduced the v2 client interface.
The Request implementation is now responsible for delegating the
HTTP request to the Volley library (previously it was the job of
the client implementation).

The Client implementation is however still responsible for
authentication requests, which is reflected through the client
interface now (previously the responsibility was exposed through
a special-case request object).

The client has also been reduced in complexity as the request
queue and worker thread spawning has been removed from it. The
Volley library provides for both features which is relied upon
now. This removes an entire layer of thread and enables a much
leaner implementation which is much easier to debug.