Refactored the `Provider` responsibility chain slightly.
Previously the `VollyProvider` was responsible to hold
and provide the an instance of the `Client` implementation.
This responsibility has now been moved a bit down in the
hierarchy, to the abstract `Provider` implementation. In
this way, custom providers, that don't necessarily provide
content, doesn't have to extend the implementation
specific `VolleyProvider` in order to get hold of the
`Client`.