Simplified session change flag management.
The flag deciding whether the session listeners should be notified
or not upon request result was previously tested in combination
with the "is auth request" flag (set for requests that manage auth
state like login, refresh etc). The two have now been split, hence,
clarifying each flags responsibility.

The session change flag was also exposed to external manipulation
in the previous implementation. This feature is no longer used by
the implementation, hence, the flag is made a strictly internal
matter of the `VolleyRequest` class.