Refactor auth error flow.
The previous implementation called EITHER the auth error
callbacks OR the generic error callbacks, but never both. This
flow has now changed to call both on authentication error.

This change enables greater flexibility when it comes to
dealing with re-auth attempts and how failures of such
should propagate to the client.