Return correct url in image loader callback.
Prior to this fix the actual request url was returned through the
callback interface. This may have changed, compared to the injected url,
depending on the provided size:

"http://my.image.url" -> "http://my.image.url/large"

After this fix the url reported back through the callback interface will
always be the exact url provided by the caller when requesting an image
to be loaded.