Refactor the initialization of the local store.
Previously the local store was initialized on a background thread and
the resulting ready-to-use store object was delivered through a callback
interface provided by the caller. This work flow turned out to be very
disruptive and caused a very complex client side implementation.

This change will initialize the memory-cache immediately on the calling
thread and return it directly on the "open()" call. The disk store is
initialized on a separate worker thread and attached to the store object
at a later point.

This adds a more convenient way of initializing and using the store, but
the caller is not notified when the disk store is prepared. There are
means to check whether the disk store is ready, though, which enables
the caller to make a decision whether or not to continue using the store
without any persistent storage.