Soften the consequences of an "Erase" request.
The previous implementation closed the store altogether on an erase
request, meaning that the memory and disk store handles were disposed.
This in turn caused some unexpected exceptions when the erase request
finished before a, say, put request.

As of this implementation the store content is erased, as the naming
suggests, but the store container itself (i.e. the file system folders)
are left intact.