untested and undocumented refactor that enables multi-instances
The approach to enable multiple instances is to allow multiple instances
of the Podio facade along with all the internal singleton/static
classes, methods and attributes so that a client (e.g. the podio app)
can create multiple independent instances of the SDK.

Main problem was to figure out how to share the session and callback
managers objects to all the classes that interact with them withouth
having them as statics.

Three things are missing for the refactor to be done:

1. make the needed changes in the Podio client so it can compile

2. test that this refactor solution actually works as intended (create
multiple instances, sign in as both a user and an app and do a couple of
API calls)

3. update the java doc of the methods and classes that have been created
and modified so it reflects the changes of the refactor