Updated date field with some contains methods
Also some of the dates in the date field was parsing utc
while in fact some of them was in the users timezone
set in the api. It’s not guaranteed ofc that the device
default timezone is in what is configured on web but
it’s at least better than assuming it’s utc when it’s not.